Fix Issue #36
update documentation of Label.TextColor to be more specific
update documentation of Lable.TextColorOverride
update UnitTests to use TextColorOverride instead of TextColor

Notes:
TextColor is used internal to set text colors when highlighting,
hovering etc from the active skin. Custom changes to this field didn't
color the texts because it was overidden internally to the default
TextColor of the active skin.
Use TextColorOverride to apply a color to the text.